+++ darts/models/forecasting/baselines.py
@@ -265,7 +265,8 @@ class NaiveMovingAverage(LocalForecastingModel):
 class NaiveEnsembleModel(EnsembleModel):
     def __init__(
         self,
-        models: List[ForecastingModel],
+        forecasting_models: List[ForecastingModel],
+        train_forecasting_models: bool = True,
         show_warnings: bool = True,
     ):
         """Naive combination model
@@ -278,8 +279,12 @@ class NaiveEnsembleModel(EnsembleModel):
 
         Parameters
         ----------
-        models
+        forecasting_models
             List of forecasting models whose predictions to ensemble
+        train_forecasting_models
+            Whether to train the `forecasting_models` from scratch. If `False`, the models are not trained when calling
+            `fit()` and `predict()` can be called directly (only supported if all the `forecasting_models` are
+            pretrained `GlobalForecastingModels`). Default: ``True``.
         show_warnings
             Whether to show warnings related to models covariates support.
 
@@ -301,11 +306,16 @@ class NaiveEnsembleModel(EnsembleModel):
                [485.16604194]])
         """
         super().__init__(
-            models=models,
-            train_num_samples=None,
+            forecasting_models=forecasting_models,
+            train_num_samples=1,
             train_samples_reduction=None,
+            train_forecasting_models=train_forecasting_models,
             show_warnings=show_warnings,
         )
+
+        # ensemble model initialised with trained global models can directly call predict()
+        if self.all_trained and not train_forecasting_models:
+            self._fit_called = True
 
     def fit(
         self,
@@ -318,13 +328,13 @@ class NaiveEnsembleModel(EnsembleModel):
             past_covariates=past_covariates,
             future_covariates=future_covariates,
         )
-        for model in self.models:
-            kwargs = dict(series=series)
-            if model.supports_past_covariates:
-                kwargs["past_covariates"] = past_covariates
-            if model.supports_future_covariates:
-                kwargs["future_covariates"] = future_covariates
-            model.fit(**kwargs)
+        if self.train_forecasting_models:
+            for model in self.forecasting_models:
+                model._fit_wrapper(
+                    series=series,
+                    past_covariates=past_covariates,
+                    future_covariates=future_covariates,
+                )
 
         return self
 
@@ -344,9 +354,6 @@ class NaiveEnsembleModel(EnsembleModel):
             logger,
         )
 
-        if series is None:
-            series = self.training_series
-
         if isinstance(predictions, Sequence):
             return [
                 self._target_average(p, ts)
@@ -363,7 +370,7 @@ class NaiveEnsembleModel(EnsembleModel):
 
     def _target_average(self, prediction: TimeSeries, series: TimeSeries) -> TimeSeries:
         """Average across the components, keep n_samples, rename components"""
-        n_forecasting_models = len(self.models)
+        n_forecasting_models = len(self.forecasting_models)
         n_components = series.n_components
         prediction_values = prediction.all_values(copy=False)
         target_values = np.zeros(
@@ -391,12 +398,12 @@ class NaiveEnsembleModel(EnsembleModel):
     def _params_average(self, prediction: TimeSeries, series: TimeSeries) -> TimeSeries:
         """Average across the components after grouping by likelihood parameter, rename components"""
         # str or torch Likelihood
-        likelihood = getattr(self.models[0], "likelihood")
+        likelihood = getattr(self.forecasting_models[0], "likelihood")
         if isinstance(likelihood, str):
-            likelihood_n_params = self.models[0].num_parameters
+            likelihood_n_params = self.forecasting_models[0].num_parameters
         else:  # Likelihood
             likelihood_n_params = likelihood.num_parameters
-        n_forecasting_models = len(self.models)
+        n_forecasting_models = len(self.forecasting_models)
         n_components = series.n_components
         # aggregate across predictions [model1_param0, model1_param1, ..., modeln_param0, modeln_param1]
         prediction_values = prediction.values(copy=False)